Refactor srv/api/chat/message.ts: split monolithic API into focused modules
- messageTypes.ts (52 lines): Type definitions and request validators
- communicationUtils.ts (22 lines): WebSocket messaging utilities
- messageUtils.ts (170 lines): Message creation, sanitization, and data restoration
- messageProcessing.ts (98 lines): User message creation and bot membership management
- responseHandlers.ts (184 lines): Response handling for authenticated and guest users
- messageHandlers.ts (67 lines): Basic message CRUD operations
- generationHandlers.ts (410 lines): Complex AI generation pipeline and streaming
- message.ts (7 lines): Clean re-export interface for backward compatibility

Original file was 1,122 lines handling too many responsibilities.
Now organized into single-purpose modules following domain boundaries.
Build and start tests pass, maintains API compatibility.
